requirements.txt
sentinelhub
matplotlib
numpy
pandas
requests
scikit-learn
scipy
tqdm
cdsapi
xarray
netCDF4
cdsapi
ecmwf-datastores-client
torchgeo
dask
opencv-python
tensorboard
tensorboardX